[MOVE/#1] MainActivity 위치 이동
