+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TerningAndroid"
         tools:targetApi="31">
-        <activity
-            android:name=".MainActivity"
-            android:exported="true"
-            android:screenOrientation="portrait"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.TerningAndroid">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
